Split FBXGeometry class to separate modules
